rename helper module in tests
